+++ src/components/Authenticated/Image.vue
@@ -1,7 +1,7 @@
 <template>
 	<v-row justify='center' align='center' no-gutters>
 		<v-col cols='auto' class='parent' width>
-			<v-img v-if='!piOnline' src='@/assets/cached.png' class='cached' contain />
+			<!-- <v-img v-if='!piOnline' src='@/assets/cached.png' class='cached' contain /> -->
 			<v-img :src='image' contain class='image-border main-image' width='650px'/>
 			<v-img src='@/assets/watermark.png' class='watermark' contain />
 		</v-col>
@@ -11,7 +11,7 @@
 <script lang='ts'>
 
 import Vue from 'vue';
-import { ImageModule, PiStatusModule } from '@/store';
+import { ImageModule } from '@/store';
 
 export default Vue.extend({
 	name: 'image-component',
@@ -19,9 +19,9 @@ export default Vue.extend({
 		image () : string {
 			return ImageModule.image;
 		},
-		piOnline () :boolean {
-			return PiStatusModule.online;
-		},
+		// piOnline () :boolean {
+		// 	return PiStatusModule.online;
+		// },
 	},
 
 	data: () => ({
@@ -55,13 +55,13 @@ export default Vue.extend({
 	z-index: 2;
 }
 
-.cached {
+/* .cached {
 	position: absolute;
 	top: 0;
 	left: 0;
 	max-width: 50%;
 	z-index: 2;
-}
+} */
 
 .main-image {
 	z-index: 1;